Prepare for druid v0.3.0
- reexport druid_shell as 'shell'
- consolidate some druid_shell pub items (WindowBuilder in base,
renaming 'win_main' -> 'runloop'
- Update for removal of Coord & Point from piet
- expose updated kurbo
- update examples

+++ druid-shell/src/window.rs
@@ -164,9 +164,9 @@ pub enum Cursor {
 #[derive(Debug)]
 pub struct ScrollEvent {
     /// The scroll wheel’s horizontal delta.
-    pub dx: f32,
+    pub dx: f64,
     /// The scroll wheel’s vertical delta.
-    pub dy: f32,
+    pub dy: f64,
     /// Modifiers, as in raw WM message
     pub mods: u32,
 }
